+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-    "target": "ES2015",
+    "target": "ES2017",
     "useDefineForClassFields": true,
     "lib": ["DOM", "ES2017", "ES2019.Array"],
     "strict": true,
